Create preprocessing
involves converting events into numerical sequences, often using techniques like one-hot encoding or integer encoding. These sequences are then padded to ensure they all have the same length. For handcrafted features like 'weekend', 'part_day', and 'nview', they are used as is but may be normalized or standardized if necessary (using methods like MinMaxScaler or StandardScaler). The input to the model consists of symbolized events in the form of sequence data and handcrafted features as a separate vector.